refactor(mocha): split lib.rs into lib, scanner, helpers, and tests modules
The 1727-line crates/mocha/src/lib.rs is split into four modules following
the same shape as the storybook and cypress refactors. All existing unit
tests still pass.
